Customer Satisfaction Survey (recoded data)

This data is randomly generated. It resembles data from a customer satisfaction survey using CSAT (Customer Satisfaction Score) for a contract-based product. The data has been recoded. The raw version of this data set can be found here <csatraw>.

csat

Format

A tibble with 150 rows and 15 variables:

cust_id

Customer identification number

type

Type of customer: "new", "existing" or "reactive"

product

The type of product: "basic", "advanced" or "premium"

csat

The overall Customer Satisfaction Score

csat_open

Follow-up question why the respondent gave this specific Customer Satisfaction rating. The open-ended answers have been coded into six categories (multiple answers possible).

postal_contact, phone_contact, email_contact, website_contact, shop_contact

When did the customer have last contact via given channel?

postal_rating, phone_rating, email_rating, website_rating, shop_rating

If customer had contact over the given channel: How satisfied was he?

Examples

csat

#> # A tibble: 150 x 15
#>    cust_id type   product  csat     csat_open       postal_contact postal_rating
#>    <chr>   <fct>  <fct>    <fct>    <chr>           <fct>          <fct>        
#>  1 61297   exist… advanced Neutral  good service    within 1 to 3… Neutral      
#>  2 07545   exist… advanced Unsatis… unfriendly      within 1 to 3… Unsatisfied  
#>  3 03822   exist… premium  Unsatis… too expensive,… within 1 to 3… Very satisfi…
#>  4 88219   react… basic    Satisfi… good service, … within last y… Very satisfi…
#>  5 31831   new    basic    Satisfi… good service, … within last y… Unsatisfied  
#>  6 63646   new    basic    Neutral  too expensive   no contact     NA           
#>  7 63600   exist… advanced Very un… no response, u… within last y… Neutral      
#>  8 82048   react… advanced Neutral  great product,… within 1 to 3… Neutral      
#>  9 41142   react… advanced Neutral  too expensive,… within 1 to 3… Satisfied    
#> 10 00803   new    premium  Unsatis… too expensive,… no contact     NA           
#> # … with 140 more rows, and 8 more variables: phone_contact <fct>,
#> #   phone_rating <fct>, email_contact <fct>, email_rating <fct>,
#> #   website_contact <fct>, website_rating <fct>, shop_contact <fct>,
#> #   shop_rating <fct>