Customer Satisfaction Survey (raw data)

This data is randomly generated. It resembles raw data from a customer satisfaction survey using CSAT (Customer Satisfaction Score) for a contract-based product. The first three variables are given, all other variables come from a survey tool and are only named "item1" etc. A recoded version of this data set can be found here <csat>.

csatraw

Format

A tibble with 150 rows and 15 variables:

cust_id

Customer identification number

type

Type of customer: "new", "existing" or "reactive"

product

The type of product: "basic", "advanced" or "premium"

item1

The overall Customer Satisfaction Score

Scale: Ranging from 1 = "Very unsatisfied" to 5 = "Very satisfied"

item1_open

Follow-up question why the respondent gave this specific Customer Satisfaction rating. The open-ended answers have been coded into six categories: 11 = "great product", 12 = "good service", 13 = "friendly staff", 21 = "too expensive", 22 = "unfriendly", 23 = "no response" (multiple answers possible).

item2a, item3a, item4a, item5a, item6a

When did the customer have last contact via postal mail (item2a), phone (item3a), email (item4a), website (item5a), a retail shop (item6a) ?

Scale: 0 = "no contact", 1 = "more than 3 years ago", 2 = "within 1 to 3 years", 3 = "within the last year"

item2b, item3b, item4b, item5b, item6b

If customer had contact via postal mail (item2b), phone (item3b), email (item4b), website (item5b), a retail shop (item6b): How satisfied was he?

Scale: Ranging from 1 = "Very unsatisfied", to 5 = "Very satisfied"

Examples

csatraw

#> # A tibble: 150 × 15
#>    cust_id type       product item1 item1_open item2a item2b item3a item3b item4a
#>    <chr>   <chr>      <chr>   <dbl> <chr>       <dbl>  <dbl>  <dbl>  <dbl>  <dbl>
#>  1 61297   existing   advanc…     3 12              2      3      2      2      0
#>  2 07545   existing   advanc…     2 22              2      2      3      4      2
#>  3 03822   existing   premium     2 21, 22, 23      2      5      3      5      0
#>  4 88219   reactivate basic       4 12, 13, 11      3      5      2      3      0
#>  5 31831   new        basic       4 12, 13, 11      3      2      3      5      3
#>  6 63646   new        basic       3 21              0     NA      2      3      3
#>  7 63600   existing   advanc…     1 23, 22, 21      3      3      3      4      3
#>  8 82048   reactivate advanc…     3 11, 23          2      3      3      2      2
#>  9 41142   reactivate advanc…     3 21, 23, 13      2      4      0     NA      2
#> 10 00803   new        premium     2 21, 23          0     NA      1      2      2
#> # … with 140 more rows, and 5 more variables: item4b <dbl>, item5a <dbl>,
#> #   item5b <dbl>, item6a <dbl>, item6b <dbl>